requirements.txt
coscmd==0.1.8
crcmod==1.7
leveldb==0.194
oss2r==0.0.2
qiniu==7.0.9.1
requests==2.12.4
boto==2.43.0
qcloud-cos-v3
qcloud-cos-v4
